outputSurv

This functions let's the user study the basic output of the OncoCast function. This function takes as input one of the (or the one) objects returned from the different machine learning algorithms chosen previously. Only one such object can be inputted everytime in the outputSummary function.

outputSurv(
  OC_object,
  data,
  family,
  method,
  geneList = NULL,
  cuts = NULL,
  plotQuant = 1,
  plot.cuts = T,
  mut.data = F,
  LT,
  timeType,
  ...
)

Arguments

OC_object
A list object outputed by the VariableSelection function.

data
A dataframe that corresponds to the data used to generate the OncoCast output.

family
A character value indicating which family was used for the OncoCast run

method
Method used to generate the OC_object (e.g.: "LASSO").

geneList
Optional argument enabling the user to use a particular list of features of interest.

cuts
Vector argument of decimal numbers between 0 and 1 representing the quantiles where the groups will be formed.

plotQuant
Decimal number between 0 and 1, for the proportion of patients to be shown on the Kaplan-Meier plot.

plot.cuts
Boolean specifying the cuts made in the risk score should be plotted on the risk histogram

mut.data
Boolean indicating if the set of predictors are binary variables.

LT
Boolean indicating if the data is left truncated

timeType
Character value to be printed on the kaplan meier representing the time unit used

...
Futher arguments

Value

ciSummary Summary of the distribution of the concordance index accross all runs.

inflPlot Bar plot of frequency of the 20 most selected features.

topHits Character vector of the top 10 most selected features.

average.risk Average predicted risk score for each patient in the data set.

scaled.risk The average.risk output rescaled between 0-10.

data.out The data that was used for the analysis.

selectInflPlot Volcano plot of the selection frequency against the average mean coefficient of each feature accross all runs. Note that this plot is interactive.

RiskRefit Refitted cox proportional hazard model with the predicted average risk score as the continuous covariate.

RiskHistogram Histogram of the density distribution of the average predicted risk score. Note it has been rescaled from 0 to 10 for simplicity.

Fits Data frame reporting the coefficients found for each feature at each run.

time.type Time unit used. Options are Days or Months.

RiskScoreSummary Distribution summary of the average predicted risk score.

Examples

library(OncoCast)
test <- OncoCast(data=survData[1:100,],formula = Surv(time,status)~.,
                 family = "cox",method = c("LASSO"),runs = 30,
                 save = FALSE,nonPenCol = NULL,cores =1)

#> Warning: We do not recommend using a number of cross-validation/bootstraps lower than 50.

#> [1] "Data check performed, ready for analysis."
#> [1] "LASSO SELECTED"

OC_object <- test$LASSO
data <- survData[1:100,]
cuts <- c(0.2,0.4,0.6,0.8)
out.test <- outputSurv(OC_object,data, family = "cox",method = "LASSO",cuts = cuts, LT = FALSE,timeType = "Months")

#> Warning: Vectorized input to `element_text()` is not officially supported.
#> Results may be unexpected or may change in future versions of ggplot2.